requirements.txt
fake-useragent
tls-client
typing-extensions
httpx
tabulate
python-dotenv 
requests
pymongo
fastapi
uvicorn[standard]
pyyaml
